docs(json): NgModules update (#69)
* created file for json-formatter.module and added forRoot() method

* updated json-formatter docs with setup

* updated README.md with setup
