+++ src/platform/json-formatter/json-formatter.module.ts
@@ -1,0 +1,27 @@
+import { NgModule, ModuleWithProviders } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { MdIconModule } from '@angular2-material/icon';
+
+import { TdJsonFormatterComponent } from './json-formatter.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    MdIconModule.forRoot(),
+  ],
+  declarations: [
+    TdJsonFormatterComponent,
+  ],
+  exports: [
+    TdJsonFormatterComponent,
+  ],
+})
+export class CovalentJsonFormatterModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: CovalentJsonFormatterModule,
+      providers: [],
+    };
+  }
+}
